requirements.txt
# FastAPI 和相关依赖
fastapi
uvicorn
starlette
python-multipart
# 数据库相关
sqlalchemy
pandas
openpyxl
# 图像处理
pillow
# 密码加密
bcrypt
# 环境变量加载
python-dotenv
# HTTP 请求
requests
# 其他工具
pydantic
